AI/speech/requirement.txt
fastapi[all]
transformers
boto3
openai
python-dotenv
opentelemetry-distro==0.46b0